+++ CONTRIBUTING.md
@@ -8,9 +8,7 @@ Some of the resources to look at if you're interested in contributing:
 * Look at the [contributing guide in our documentation](https://brynpickering.github.io/python_boilerplate/contributing)
 
 ## Licensing
-
-By contributing to python_boilerplate, i.e. through opening a pull request, you represent that your contributions are your own original work and that you have the right to license them, and you agree that your contributions are licensed under the python_boilerplate.
-
+By contributing to python_boilerplate, i.e. through opening a pull request, you represent that your contributions are your own original work and that you have the right to license them, and you agree that your contributions are licensed under the MIT license.
 ## Reporting bugs and requesting features
 
 You can open an issue on GitHub to report bugs or request new python_boilerplate features.
