+++ conda.recipe/meta.yaml
@@ -56,4 +56,4 @@ about:
   doc_url: {{ pyproject.project.urls.documentation }}
   summary: {{ pyproject.project.description }}
   license: {{ pyproject.project.license.text }}
-  license_file: LICENSE 
+  license_file: LICENSE
