+++ tests/test_cli.py
@@ -1,0 +1,16 @@
+"""Tests for `python_boilerplate` CLI."""
+
+from click.testing import CliRunner
+
+from python_boilerplate import cli
+
+
+def test_command_line_interface():
+    """Test the CLI."""
+    runner = CliRunner()
+    result = runner.invoke(cli.cli)
+    assert result.exit_code == 0
+    assert "python_boilerplate.cli.cli" in result.output
+    help_result = runner.invoke(cli.cli, ["--help"])
+    assert help_result.exit_code == 0
+    assert "--help  Show this message and exit." in help_result.output
