+++ tests/test_core.py
@@ -3,9 +3,8 @@
 """Tests for `python_boilerplate` package."""
 
 import pytest
-from click.testing import CliRunner
 
-from python_boilerplate import cli, python_boilerplate
+from python_boilerplate import core
 
 
 @pytest.fixture
@@ -20,16 +19,7 @@ def response():
 
 def test_content(response):
     """Sample pytest test function with the pytest fixture as an argument."""
-    print(python_boilerplate.__file__)
+    print(core.__file__)
     # assert 'GitHub' in BeautifulSoup(response.content).title.string
 
-
-def test_command_line_interface():
-    """Test the CLI."""
-    runner = CliRunner()
-    result = runner.invoke(cli.cli)
-    assert result.exit_code == 0
-    assert "python_boilerplate.cli.cli" in result.output
-    help_result = runner.invoke(cli.cli, ["--help"])
-    assert help_result.exit_code == 0
-    assert "--help  Show this message and exit." in help_result.output
+    #!/usr/bin/env python
